fix(release): publishing jobs fail when new commits are pushed
When new commits are pushed to the release branch, the release job will not upload the artifact but the publisher jobs will still execute (and fail, because there is no artifact).

This change adds a job condition that will only execute the publishing jobs if we are indeed releasing the latest commit. Otherwise, the publishers will be skipped and the workflow will not fail.

Fixes #723