fix: rename license files to correct SPDX ids (#980)
Hi Projen folk, I'm a part of the AWS OSPO and I noticed that the license filenames were not using SPDX codes. My understanding is that these are licenses that someone generating project configuration might use as the license for their output. 

Some of them also seemed like odd choices, either due to being unusual or being old dated versions of licenses that folk probably do not use anymore.

I don't know what the effect of doing this is - I'm not a clueful projen user. 

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.