fix: tests under lib/ are included in npm tarballs (#992)
Fixes #955 

I don't think it's possible to fix #955 without disabling the automatic sorting behavior in `IgnoreFile`, so I tried changing the implementation to see if it makes sense. The main difference is that the underlying data structure is an array, and not a pair of sets (current implementation) or an object (implementation from 8+ months ago).

One feature I added is that `IgnoreFile` will eliminate existing patterns that are "overridden" by the patterns you add. So if you add `/foo/`, then not only will `!/foo/` be removed, but also any other patterns starting with `!/foo/` such as `!/foo/bar/baz` or `!/foo/**/*.ts`. Likewise, adding `!/foo/` will remove any patterns starting with `/foo/`. This kind of logic only applies to directories (we don't want `/foo` to overwrite `!/foo.txt` because `foo` might be the name of a file with no extension).

I also deprecated the `npmIgnore` and `gitIgnore` fields in `NodeProject`. Since any patterns added via these fields would be applied before patterns added in subclasses such as `TypeScriptProject`, this will lead to a [confusing developer experience](https://github.com/projen/projen/issues/453) under the new order-dependent implementation. Instead, users should use `project.addPackageIgnore()` and `project.addGitIgnore()`. I have 1-2 ideas for resolving this, but they each add a bit of complexity so I might take them to a separate issue.

One downside of disabling sorting is that the generated output is now dependent on the order in which patterns are added, so refactorings could cause npmignore files to get randomly reshuffled.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.