fix: unclear property for detecting failed releases (#1035)
Previously, the property for creating an issue on failed releases was called `failureIssue`.

Even though this property is part of the `ReleaseOptions` struct, its usage is unclear because these properties are embedded inside the `NodeProjectOptions`, hence loosing its *release* context.

Rename to `releaseFailureIssue` to make it clear it relates to the release process.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.