fix: projen incompatible with node 10.x (#1063)
A recent change updated the minimum node version of projen itself to 14.x. This implies that it is now impossible to use it in environments with older node, including any existing workflows.

In the meantime, we set the min version back to 10.x.

NOTE: This change also relaxes engine checks during projen new. This technically means that users are able to create a new project within a non-compatible environment, but since engine checks are performed all the time anyway, they won't be able to go a long way without a compatible node version. The reason this was needed is because our "new" tests use the latest version of node when they test external projects and version 0.28.0 now requires node 14.x... Anyway - a bit of a mess.