feat: addOverride can override array elements (#1067)
Previously, it was not possible to override specific array elements through the `addOverride` escape hatch, even though this is possible in some other CDK libraries (like the AWS CDK). This adds this feature back.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.