feat(eslint): import aliasing (#1048)
allow specification of configuration to remove lint errors for projects that utilise import aliasing.

e.g `import { bar } from '@src/foo'`


## Notes
I'm not entirely happy with the API here. I'd have preferred to take an option for `alias` as an object. However, when trying to compile I got jsii errors saying map types were unsupported

```typescript
readonly alias?: {
  map: [string, string][];
  extensions: string[];
};
```

If there is a way to support this kind of API I believe it would be more intuitive for the consumer than having different options for the map and the extensions as submitted in this PR.

```typescript
  /**
   * Enable import alias for module paths
   * @default undefined
   */
  readonly aliasMap?: { [key: string]: string; };

  /**
   * Enable import alias for module paths
   * @default undefined
   */
  readonly aliasExtensions?: string[];
```

### Possible future task

All TypeScript projects need to configure import aliasing in 3 places
- tsconfig
- eslint config
- jest config


tsconfig and jest already exposed ways to set this but eslint did not. Given that projen generates all three of these files, there is opportunity to define the mapping in one place and inject it into all three config files. I'd have to do some digging into project structure to figure out how this might work but happy to help with this if you think it's worth having

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.