fix(new): Unable to use external project from many package specs (#1094)
Fixes #879 
(The original issue was (I think) fixed with https://github.com/projen/projen/pull/1077, this fixes the other issue brought up regarding `npx projen new --from cdk-appsync-project@latest`)

Previously, one was unable to use `--from`  with a spec that was not of the form `package` or `package@version`. Meaning you could not install from a tarball or one of the other many sources, even a dist tag (e.g. `package@latest`).

This change makes use of the results of the `npm install` command to figure out the actual `package@semanticVersion` so that we can use it to parse the actual package name.
Also, changed `yarn` -> `npm` in a few comments to reflect the changes from my last PR.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.