fix: require node.js ^12.7.0 instead of ^12.13.0
The latest projen release bumped the minimum node.js version requirement to ^12.13.0 but this is technically a breaking change and blocks upgrades for projects that execute workflows against 12.7.0 (such as `constructs`).

This change reverts the minimum node engine version to 12.7.0 while still running our workflows against 12.13.0, which is required by `jest`. This is not perfectly correct as it means projen may depend on functionality that does not exist in 12.7.0 but its pragmatic and will cause less grief.