chore: fix projen workflow tokens (#1134)
Currently in the projen repo, dependency upgrade PRs generated by the `upgrade-main.yml` workflow do not get automatically approved because the PRs are generated using the GITHUB_TOKEN, which is the same token that is running the `auto-approve.yml` workflow. To fix this, we set the upgrade dependency workflow to use a different token for creating the dependency upgrade PRs.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.