chore(project): default outdir to /tmp in tests (#1157)
When running in the context of a jest test, default `outdir` to a temporary directory instead of cwd. Otherwise, by default, projects created in tests will override the current project and that’s undesirable.
It also cleans up a lot of the boilerplate when creating projects in tests.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.