feat: standard build (#1201)
Define a common build process for all projects which includes a set of standard phases that can be used for extensibility. The idea is that components and project types will have a solid foundation for build extensions instead of ad-hoc set of tasks for each project.

The `build` task now spawns a set of standard build phase tasks:
- `default` (synthesize project - executes your projenrc).
- `pre-compile`
- `compile`
- `post-compile`
- `test`
- `package`

The `Project` base class now has a set of properties that can be used to access these tasks. For example, `project.postcompileTask` will return the `Task` that’s executed after compilation.

The method `task.lock()` can now be used to lock a task for mutations. The `build` task is now locked, which means that any modifications to it will throw an exception.

This change also includes some logging improvements and cleanups.

This is an attempt to create generic build model for all project types. We shall see if this holds water.

BREAKING CHANGE: It is now impossible to modify the `build` task. Instead, extend one of the specific build phases (`precompile`, `compile`, `post compile`, `test` and `package`). To access these tasks use `project.xxxTask`.
* The `compileBeforeTest` option in `TypeScriptProject` is not supported any more. Tests are always executed _after_ compilation.
* `projenDuringBuild` is no longer supported. Let us know if you have a use case for it that we are not aware of.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.