fix(node): bundled assets cannot be located during tests (#1200)
When test code is not executed from `lib/`, bundled assets cannot be located because they are under `lib/`. To address this we now place assets under a root `assets/` directory which has the same relative path whether code runs from `lib/` or `src/`.

BREAKING CHANGE: minor tweaks of the `LambdaFunction` and `Bundler` APIs.