fix: release.yml cannot be modified with escape hatches (#1255)
Fixes #1103

Resolves a bug where the workflows generated by the `Release` component cannot be accessed or modified using escape hatches because they are generated during the `preSynthesize()` stage of the component lifecycle.

This requires making some changes to `Publisher`, which `Release` relies upon, so that it can provide a link to the jobs it will generate even if `Publisher`'s jobs are not all available at the time construction finishes.

In addition, workflows that are not created via the `github.addWorkflow` method were not getting added to `GitHub` so they could not be discovered with `tryFindWorkflow`. This is fixed by adding a register hook within the `GithubWorkflow` constructor.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.