feat(awscdk): integration tests (#1270)
Add support for CDK integration tests.

To use, create a file with an `xyz.integ.ts` suffix under the test tree and run `projen integ:xyz:deploy` to deploy it to your development account and capture a snapshot of the synthesized output. During build, as part of the `test` phase, the snapshot will be verified.

BREAKING CHANGE: `AutoDiscover` now requires `testdir`.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.