feat(release): allow specifying npm dist-tag per branch
Adds `npmDistTag` to release branch definition which will be configured in the release workflow triggered for this branch.

Implementation notes: to support this we needed to refactor the publisher so that when its jobs are rendered, they have the branch context. This also means that now we can ensire that when a publish command is executed manually, it will be done only on the correct release branch.

The root `npmDistTag` setting applies to the default branch.

BREAKING CHANGE: `npmDistTag` only effects publishing (and supported for each release branch) setting and so it is no longer set in `package.json`.
* `workflow.addJobsLater()` is no longer supported. Use `postSynthesis` hooks and call `.addJobs()` as needed.