fix(cleanup): file pruning is too heavy handed (#1306)
Added a files.json in the projen folder to store a list of the files generated by projen on the last run. Using this file, projen can selectively clean up files on subsequent runs.

With this implementation, generated files are cleaned up regardless of whether or not they have the `PROJEN_MARKER` in them. A file is considered a candidate for deletion if it is marked as readonly and is in the files array of a project.
If files.json is not present or otherwise invalid, the previous method of cleanup is used.

fixes #1265

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.