fix(awscdk): generated lambda code not compatible with CDK v2
Extract the logic that handles CDK dependencies (and supports v1/v2 differences) into a new component called `AwsCdkDeps` and reuse across  `AwsCdkTypeScriptApp` and `AwsCdkConstructLibrary`.

Then, use it in `LambdaFunction` to determine how to render import statements in the generated function construct source code.

Change the default CDK version for `awscdk-app-ts` to `^2`.

BREAKING CHANGE: The `cdkVersion` option is now a *semver range* and not just a specific version. To enable the previous behavior, prefix with `^`.
* **awscdk:** `LambdaFunction` and `AutoDiscover` now requires `cdkDeps` so they can interact with CDK version manager.