chore: use major-pinned version of @types/node
`@types/node` versions numbers match the node runtime versions'
`major.minor`, however new releases are only created when API changes
are included in a node release... We might for example have dependencies
that require `node >= 12.22`, but as 12.21 and 12.22 did not include API
changes, `@types/node@12.20.x` is the "correct" version to use. As it is
not possible to easily determine the correct version to use, we pick up
the latest version.