fix(awscdk): issues with CDK v2 libraries and apps (take 2) (#1338)
Extract the logic that handles CDK dependencies (and supports v1/v2 differences) into a new component called `AwsCdkDeps` and reuse across  `AwsCdkTypeScriptApp` and `AwsCdkConstructLibrary`.

This is a follow up on #1327 which we had to revert due to breaking upgrade flows. The difference is that this change does NOT introduce breaking behavioral change to `cdkVersion` and dev/peer dependencies of constructs.

* Modify `LambdaFunction` to determine how to render import statements in the generated function construct source code.
* Fix the sample test code in `awscdk-app-ts` to work with the new assertions library.
* Change the default CDK version for `awscdk-app-ts` to `2.1.0`.
* Determines workflow container image based on CDK version

Fixes #1288
Fixes #1326
Fixes #1173

Co-authored-by: Christopher Rybicki <rybickic@amazon.com>
Co-authored-by: github-actions <github-actions@github.com>

BREAKING CHANGE: `LambdaFunction` and `AutoDiscover` now requires `cdkDeps` so they can interact with CDK version manager.