feat(jsii): parallelize language-specific packaging jobs
Instead of running `jsii-pacmak` within the build job, fan out to concurrent jobs, one for each language.

In the `release` workflow, this work has been moved to the per-language publishing jobs. In the `build` workflow, a new job was added for each language.

This allows us to stop using the jsii/superchain image for the primary build and to improve build times, and also reduce surface area of the environment.

TODO:

- [ ] Centralize knowledge about per-language setup steps
- [ ] Handle build check updates (it should now happen _after_ all jobs were successful)
- [ ] Decide what to do in `upgrade` workflows?