feat: pwd artifact (#1384)
Instead of only using the `artifactsDirectory` as the build artifact, use the entire current working directory. 

This is much more flexible since it allows downstream jobs access to the entire repo. It effectively makes the post build jobs run on the same environment as they would have if the were executed from the same job.

### Caveat 

For node projects, `node_modules` is not included in the artifact because it takes forever to compress and upload. 
If a post build job needs `node_modules`, the job is responsible for installing it, using the lock files available in the articact.

Fixes https://github.com/projen/projen/issues/1381

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.