fix: release does not bump (#1408)
The approach proposed in https://github.com/projen/projen/pull/1384 fails the release workflow. The problem is that the `release` task will run `unbump`, reverting the `package.json` version to `0.0.0`.

Instead, we go back to the standard artifacts mechanism of uploading the contents of the `dist` directory. In jsii projects, the dist directory will contain the entire built repository, including the bump. This is done by resetting the default `package` task to simply copy over the repo into `dist`.

When we download the artifact, we have the built repository in the `dist` directory, and we run packaging and publishing from this directory, instead of the root. 

Note that for non-jsii projects, the artifacts will simply be the tarballs, so they will not have access to the repository.  

> This is all still too funky, and should be revisited when we think about the new/enhanced modeling. 

Also, intentionally flip to dry run releases so I can inspect the workflow and artifacts before possibly publishing another bad version. 

Fixes https://github.com/projen/projen/issues/1407

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.