fix(workflows): status check update is not working (#1383)
Use a Personal Access Token to push changes and create pull requests from the `build` and `upgrade` workflows. This implies that GitHub will trigger workflows as a result, and removes the need to manually create a "fake" status check update. It also simplifies the upgrade workflow because now it doesn't need to re-implement "build", which just got fancier in JSII projects (packaging tasks are executed in parallel).

Since this cannot be done with a normal GITHUB_TOKEN, and combined with the limitations of this token in upgrade workflows (unable to push workflow updates), we decided that from now, projen _requires_ a dedicated personal access token (PAT) in the repo.

By default, the name of this secret is `PROJEN_GITHUB_TOKEN` (which is the name most projects use today). If this secret does not exist in the repository, at the moment it will result in a poor experience. We will consider improving this experience in the future.

Also fixes #1362 by adding a separate job that only executes when build is triggered by forks, and prints a nice error with the diff.

BREAKING CHANGE: A secret called `PROJEN_GITHUB_TOKEN` with `workflows`, `packages` and `repo` scopes is now **required** for all projects. As much as this is a bit of a burden, most projects already have such a token because otherwise projen upgrade workflows cannot really function.