fix(workflows): mutation check did not consider unstaged changes (#1417)
We accidentally used `--staged` in the `git diff` call that checked for mutations without calling `git add .` before.

This change consolidates the mutation check into the patch upload "action" so this is done in a single place.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.