chore: RFC 1365: Project creation prompts (#1365)
There are no open issues on this currently, but a relevant one is #613.

Requesting feedback on the proposal for an interactive experience in `projen new`.

[Rendered Version](https://github.com/MarkMcCulloh/projen/blob/prompts-rfc/rfcs/project-creation-prompt.md)

Not sure what to do about the PR title starting with `chore:`, too bad there's not an `rfc` type that I'm aware of.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.