feat: extend prettier support (#1367)
allow to create a prettierrc with all config values aswell as the corresponding ignore file
it also supports overrides

I'm quite sure you already know prettier as a tool, it formats code. The initial support was already implemented but a lot of teams use prettier options for their repos. These can now be set and will be written into `prettierrc.json`. Also it allows you to ignore certain files with the prettierignore. I build a sample repo to verify this but also wrote some tests to verify that certain configs work and the corresponding files get created. 

fixes #1318

Co-authored-by: @Crapatrick <patrick@domnick.dev>

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.