fix(build): build should fail if mutations were found (#1428)
Fixes #1423 and fixes #1422 by failing the build job when a mutation was found and allowing the `self-mutation` and the `anti-tamper` jobs to execute even if the `build` job failed.

Here is an [example](https://github.com/projen/projen/runs/4664625019).

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.