fix(node): lockfile not updated after "*" deps are resolved (#1451)
#1435 introduced an optimization to only run `yarn/npm install` if package.json has changed. This uncovered a latent issue which already existed.

When a dependency version is not specified, projen initially synthesizes a `package.json` file with a `*` dependency. Then invokes `yarn install`, resolves the actual version from the installed module and updates `package.json` with a caret version requirement. This mimicks the same behavior as `yarn add` or `npm install`.

But now we need to re-execute "install" in order to update the package lock file.

Fixes #1443

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.