chore: make projen pretty(ier) (#1453)
Enable `prettify` for this project, and reformat all source files.

I contemplated if we should configure prettify to align a bit more closely with the previous eslint formatting rules (a glaring difference is quotes versus double-quotes), but I am thinking to just rip the band-aid and accept the defaults.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.