fix(awscdk): permission denied error for Lambda function with `prettier` option (#1466)
Prettier cannot format the generated Lambda construct file because it is a readonly file.
If `prettier` is enabled, the source code file must be be added to `.prettierignore` to exclude it from prettier.

This PR adds a `preSynthesize` step for prettier that will add all `SourceCode` files to the `.prettierignore` file.

Fixes #1457.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.