new code to handle older git that does not do 'init -b' command (#1485)
This PR supercedes https://github.com/projen/projen/pull/1437.  Since all changes were to one file and the main branch had undergone significant changes while this was in flight, I close 1437 and created this new PR off the new main to keep the change set extremely understandable and easy (and so I didn't mess up any merge corrections in code I am not familiar with).

git prior to v2.28 does not directly support the "init -b" option, requiring the former commands in that scenario.  Ubuntu 20.04 still uses v2.25 and is a very common distro version since it is LTS so this fix is quite applicable.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.