fix(cleanup): Only clean up orphaned files (#1494)
To avoid unnecessary thrashing of projen managed files, we can skip cleaning up files that we know we're just going to be creating again during synthesis. This means that changed files will simply be overwritten during synthesis. Even better, unchanged files won't be written at all.
With this PR, an orphaned file is one that is present in `files.json` (from the previous synthesis) but not one of the files that are about to be synthesized.

Included some refactoring in `cleanup.ts` to make the control flow more (hopefully) clear with this additional logic.

fixes #1484 

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.