chore: resynth after compiling projen (#1516)
This fixes a feedback loop where contributors add some change to projen (that happens to affect projen's own configuration files), and running `build` will compile and run all tests but not necessarily run update projen's own config because the `compile` task runs after `default`.

This is not totally fool-proof, since it's possible the change to projen could affect the actual build steps and thus require a full build again, but it might be possible to make this smarter in the future with https://github.com/projen/projen/issues/1515.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.