fix: disabled github components throws error (#1540)
If `github: false` is configured the `NodeProject` throwing an error, when there is auto approval configured.

The error occured with an external projen project type, which has presets for github, but should just be disabled on some projects. I would like to keep the configuration and just disable the github feature.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.