feat: customise build workflow trigger in NodeProject (#1562)
This PR is about allowing the build workflow trigger to be customised. Currently it defaults to `pull_request` and `workflow_dispatch` triggers and I would like to modify it with something else if possible.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.