feat(github): allow JobMatrix to accept values other than string (#1583)
Adding this feature to allow setting the JobMetrix values to other primitive data types rather than string. For instance, I want to set the session length for AWS assume role github action which is a number. This is comming as a value from matrix.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.