fix(upgrade-dependencies): upgrade task cannot be overwritten (#1602)
Small change that moves the `createTask` step of `UpgradeDependencies` into the constructor, so that it is added during project instantiation. This allows the task to be overwritten with `project.removeTask()` per the error message:
```
Error: A task with the name upgrade already exists. To override it, call removeTask first and then create the new task.
```
This previously did not work, as the error was actually being thrown by the `UpgradeDependencies.createTask()` function during the pre-synth step. I considered instead adding error handling during this step, but could not decide whether silently not creating a task if one already existed or throwing an error with a clearer message would be a better behaviour, and decided on this approach instead.

Also includes a couple of non-breaking typo fixes.

fixes #1129 

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.