fix(node-packages): pin peer dependencies with undefined version (#1609)
fixes #1611 

Peer dependencies added without a fixed version are being added to devDependencies in package.json as `name@^version` even with `pinnedDevDependency=true`, which is throwing a JSII warning. This change adds another pass through the peer dependencies in the post-synth step to pin those peer dependencies who's version is not known during presynth. While this causes a degree of code duplication, I decided against replacing the pre-synth peerDependency pass as it minimises the impact of this change on existing workflows.

This change is tested both in mocked unit tests and through yarn link to a local project, where it displayed the expected behaviour, including correctly respecting any existing versioning in `package.json`.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.