fix(upgrade-task): Fixed upgrade task to include line ending in package.json (#1650)
If the package manager is set to npm, the `npm install` command adds a new line to the end of package.json. This is a known issue: https://github.com/npm/npm/issues/18545
All of the other tasks do not do this currently, and this causes the upgrade workflow to fail for the first time and self mutation fixes that. To avoid that, added this new line to package.json anyway.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.