chore: disable stale workflow by default (#1689)
After talking with others and reviewing its impact on contributors of different GitHub projects, it feels like enabling the stale workflow may not be the right default for all projects. This PR changes the default to not adding it, so it's now opt-in.

BREAKING CHANGE: `stale.yml` GitHub workflow is no longer generated as the default behavior in projects. To keep generating it, specify `stale: true` or initialize a `github.Stale` component directly.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.